Added copyright and javadoc.
