--- src/main/java/org/craftercms/studio/api/v1/asset/processing/AssetProcessorFactory.java
@@ -1,9 +1,0 @@
-package org.craftercms.studio.api.v1.asset.processing;
-
-import org.craftercms.studio.api.v1.exception.AssetProcessingConfigurationException;
-
-public interface AssetProcessorFactory {
-
-    AssetProcessor getProcessor(ProcessorConfiguration config) throws AssetProcessingConfigurationException;
-
-}
