--- src/main/java/org/craftercms/studio/api/v1/asset/processing/AssetProcessorPipelineFactory.java
@@ -1,7 +1,0 @@
-package org.craftercms.studio.api.v1.asset.processing;
-
-public interface AssetProcessorPipelineFactory {
-
-    AssetProcessorPipeline getPipeline(ProcessorPipelineConfiguration config);
-
-}
