+++ src/main/java/org/craftercms/studio/impl/v1/asset/processing/AbstractAssetProcessor.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (C) 2007-2018 Crafter Software Corporation. All rights reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package org.craftercms.studio.impl.v1.asset.processing;
 
 import java.io.IOException;
@@ -16,6 +32,11 @@ import org.craftercms.studio.api.v1.exception.AssetProcessingException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/**
+ * Base class for {@link AssetProcessor}s.
+ *
+ * @author avasquez
+ */
 public abstract class AbstractAssetProcessor implements AssetProcessor {
 
     private static final Logger logger = LoggerFactory.getLogger(AbstractAssetProcessor.class);
