+++ src/main/java/org/craftercms/studio/impl/v1/service/asset/processing/AssetProcessingServiceImpl.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (C) 2007-2018 Crafter Software Corporation. All rights reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package org.craftercms.studio.impl.v1.service.asset.processing;
 
 import java.io.IOException;
@@ -22,7 +38,7 @@ import org.craftercms.commons.lang.UrlUtils;
 import org.craftercms.studio.api.v1.asset.Asset;
 import org.craftercms.studio.api.v1.asset.processing.AssetProcessingConfigReader;
 import org.craftercms.studio.api.v1.asset.processing.AssetProcessorPipeline;
-import org.craftercms.studio.api.v1.asset.processing.AssetProcessorPipelineFactory;
+import org.craftercms.studio.api.v1.asset.processing.AssetProcessorPipelineResolver;
 import org.craftercms.studio.api.v1.asset.processing.ProcessorPipelineConfiguration;
 import org.craftercms.studio.api.v1.exception.AssetProcessingException;
 import org.craftercms.studio.api.v1.exception.ContentNotFoundException;
@@ -33,6 +49,11 @@ import org.craftercms.studio.api.v1.service.asset.processing.AssetProcessingServ
 import org.craftercms.studio.api.v1.service.content.ContentService;
 import org.springframework.beans.factory.annotation.Required;
 
+/**
+ * Default implementation of {@link AssetProcessingService}.
+ *
+ * @author avasquez
+ */
 public class AssetProcessingServiceImpl implements AssetProcessingService {
 
     private static final Logger logger = LoggerFactory.getLogger(AssetProcessingServiceImpl.class);
@@ -40,7 +61,7 @@ public class AssetProcessingServiceImpl implements AssetProcessingService {
     private String configPath;
     private ContentService contentService;
     private AssetProcessingConfigReader configReader;
-    private AssetProcessorPipelineFactory pipelineFactory;
+    private AssetProcessorPipelineResolver pipelineResolver;
 
     @Required
     public void setConfigPath(String configPath) {
@@ -58,8 +79,8 @@ public class AssetProcessingServiceImpl implements AssetProcessingService {
     }
 
     @Required
-    public void setPipelineFactory(AssetProcessorPipelineFactory pipelineFactory) {
-        this.pipelineFactory = pipelineFactory;
+    public void setPipelineResolver(AssetProcessorPipelineResolver pipelineResolver) {
+        this.pipelineResolver = pipelineResolver;
     }
 
     @Override
@@ -84,7 +105,7 @@ public class AssetProcessingServiceImpl implements AssetProcessingService {
                     Set<Asset> finalOutputs = new LinkedHashSet<>();
 
                     for (ProcessorPipelineConfiguration pipelineConfig : pipelinesConfig) {
-                        AssetProcessorPipeline pipeline = pipelineFactory.getPipeline(pipelineConfig);
+                        AssetProcessorPipeline pipeline = pipelineResolver.getPipeline(pipelineConfig);
                         List<Asset> outputs = pipeline.processAsset(pipelineConfig, input);
 
                         if (CollectionUtils.isNotEmpty(outputs)) {
